------------------------------------------ Added
Added `::img_profile` or `::ip` or `::imgp`! See your stats in glory, and share with your friends!~

Added :notepad_spiral: icon to Suggestion window
Fixed annoying issue where the player was told they were using a command when they, in fact, were not. Hopefully, it is finally dealt with.
Added the `/adventure` command! More coming soon, but they are quite a pain to adapt tbh so keep that in mind.
There now is an achievement for being #1 on the leaderboard~
More coming soon~

------------------------------------------
Changed

Change `::heal` to be able to take in numbers. Aka `::heal <amount>` will now consume the lowest grade healing potions available to you. Keep in mind, it will not count higher grades. So if you type `::heal 10` and you have 9 Lesser healing potions, it will throw an error.
Potions (only Lesser Healing Potions) now work in `::adventure`. Just click on "Use Item".